Added STT and TTS
I added STT using VOSK. I added TTS using flite. Both STT and TTS
run in separate threads, so that the computer continues to listen
to you while it is processing your request, and can start
processing the next request while it is still talking. The audio
to analyze and the words to be spoken both get inserted into queues
so that the computer is only processing one stt at a time, and is
only saying one thing at a time, in the order added to the queue.

I thought about saving the audio output to a queue instead of the
text to be spoken, so that it wouldn't have to pause while
generating the wav output for the next spoken text, but it's not
that much and a brief pause between utterances seems natural.

The path to your VOSK model is defined in a constant at the top of
the python script.

I added a shell wrapper to set up echo cancellation in pulseaudio.

I also added a requirements.txt file which I think has most of the
python packages I had to install in order to get this to work.