Put copy of indata into record queue
In the Hello_there_echo.wav file, I was just getting the same
block of audio over and over. I just realized that this is
probably because a reference to the indata array is added to the
queue, so when pulling the data from the queue, I get a reference
to the last block of indata over and over. I changed this to
record_queue.put(indata.copy()) and got my expected audio out.