Add say method
This adds a "say" method to the TestVAD object in test.py, which
simplifies the programming some. I also added a little block of
instructions about how to use the program.

In the aectest.py program, I have increased the block size which
appears to have solved the stability issues.